refactor(api): move mock server and response to src/mocks/
